Finished Grid
The grid is now done to the specification of what is due on Tuesday.
The grid can be initialized and generates all the necessary items and
entities.  I also added a toString override on the grid which allows it
to be printed easily.  On the toString output, P = player, I =
invincibility, R = room, N = ninja, D = radar, U = bullet, B = room with
briefcase.