Item functionality
Added lots of different methods and fields for items, giving them much of their usability. I have also updated the constructors for the items this affects. Items have a "turnsleft" field that will be lowered once per turn, and items where this is unneccesary like the bullet will just ignore this field