Game Engine and testapp update
First with the test app, I commented out the original code for the first milestone, it is still there if needed. The Code I added allows the player to move around the board and pick up items; the player cannot walk out of bounds, through walls, or ninjas, all that functionality is currently there.
The commented out chunks in the test app also show some of my testing of the ninja ai which is currently in progress