Updated Ninja class to have Unique Identifier
In order to move the Ninjas, they are going to need a unique way of
identification.  I added a totalNinjas static int so that we can keep
track of all the ninjas created and use that to set their unique
identifier.  This is all done in the default constructor, and can be
returned by the getUniqueId method.