Ninja Movement Check Update
I believe I have optimized the tried movement check by only having it
move the ninja if it was not tried already like he wanted us to do.  In
my testing, it hasn't gotten stuck in any loops, but please keep testing
it.  I left the counter there so we can revert to that method quickly if
this causes problems.